Dont add duplicate entries in tag key list
Taggable caching, use a cache entry to keep a list of keys (AKA cache Id's)
That belong to a given tag. Prior to this fix, on save of a tagged cache
the key was appended to the list, even when already in the list.
Therefore when used on cache that is frequently updated
(like when using for session storing), the list size would became out of control.